rdma cmds from fabric plugin
